fix(go): permissions for network graph;
